--- src/components/square/Square.jsx
@@ -1,5 +1,0 @@
-export default function Square(props) {
-    return (
-        <button className="square" onClick={props.onClick}>{props.value}</button>
-    )
-}
