+++ src/components/square/index.js
@@ -1,3 +1,5 @@
-import Square from './Square';
-
-export default Square;
+export function Square(props) {
+    return (
+        <button className="square" onClick={props.onClick}>{props.value}</button>
+    )
+}
