+++ src/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './components/app/App';
+import { App } from './components/app';
 
 ReactDOM.render(
-    <App/>,
+    <App />,
     document.getElementById('root')
 );
