+++ src/redux/store.js
@@ -1,0 +1,6 @@
+import { rootReducer } from './reducers';
+import { createStore } from 'redux';
+
+const store = createStore(rootReducer);
+
+export { store };
